+++ packages/groveman_sentry/test/src/mock_hub.dart
@@ -8,6 +8,9 @@ import 'package:sentry/src/profiling.dart';
 class MockHub implements Hub {
   List<AddBreadcrumbCall> addBreadcrumbCalls = [];
   List<EventCall> eventCalls = [];
+  final MockScope _scope = MockScope();
+
+  MockScope get mockScope => _scope;
 
   @override
   Future<void> addBreadcrumb(Breadcrumb crumb, {dynamic hint}) async =>
@@ -23,7 +26,9 @@ class MockHub implements Hub {
   Future<void> close() async {}
 
   @override
-  void configureScope(void Function(Scope) callback) {}
+  void configureScope(void Function(Scope) callback) {
+    callback(_scope);
+  }
 
   @override
   bool get isEnabled => throw UnimplementedError();
@@ -115,7 +120,7 @@ class MockHub implements Hub {
   SentryProfilerFactory? profilerFactory;
 
   @override
-  Scope get scope => throw UnimplementedError();
+  Scope get scope => _scope;
 
   @override
   Future<SentryId> captureFeedback(
@@ -135,17 +140,22 @@ class MockHub implements Hub {
   void generateNewTrace() {}
 
   @override
-  Future<SentryId> captureException(dynamic throwable,
-      {Object? stackTrace,
-      Hint? hint,
-      SentryMessage? message,
-      ScopeCallback? withScope}) {
+  Future<SentryId> captureException(
+    dynamic throwable, {
+    Object? stackTrace,
+    Hint? hint,
+    SentryMessage? message,
+    ScopeCallback? withScope,
+  }) {
     throw UnimplementedError();
   }
 
   @override
-  Future<SentryId> captureTransaction(SentryTransaction transaction,
-      {SentryTraceContextHeader? traceContext, Hint? hint}) {
+  Future<SentryId> captureTransaction(
+    SentryTransaction transaction, {
+    SentryTraceContextHeader? traceContext,
+    Hint? hint,
+  }) {
     throw UnimplementedError();
   }
 
@@ -157,6 +167,107 @@ class MockHub implements Hub {
   @override
   void setAttributes(Map<String, SentryAttribute> attributes) {
     throw UnimplementedError();
+  }
+
+  @override
+  Future<void> captureMetric(SentryMetric metric) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> captureSpan(SentrySpanV2 span) {
+    throw UnimplementedError();
+  }
+
+  @override
+  RecordingSentrySpanV2? getActiveSpan() {
+    throw UnimplementedError();
+  }
+
+  @override
+  SentrySpanV2 startIdleSpan(
+    String name, {
+    Duration idleTimeout = const Duration(seconds: 3),
+    Duration finalTimeout = const Duration(seconds: 30),
+    bool trimIdleSpanEndTimestamp = true,
+    Map<String, SentryAttribute>? attributes,
+    DateTime? startTimestamp,
+  }) {
+    throw UnimplementedError();
+  }
+
+  @override
+  SentrySpanV2 startInactiveSpan(
+    String name, {
+    Map<String, SentryAttribute>? attributes,
+    SentrySpanV2? parentSpan = const UnsetSentrySpanV2(),
+    DateTime? startTimestamp,
+  }) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<T> startSpan<T>(
+    String name,
+    Future<T> Function(SentrySpanV2 span) callback, {
+    Map<String, SentryAttribute>? attributes,
+    SentrySpanV2? parentSpan = const UnsetSentrySpanV2(),
+    DateTime? startTimestamp,
+  }) {
+    throw UnimplementedError();
+  }
+
+  @override
+  T startSpanSync<T>(
+    String name,
+    T Function(SentrySpanV2 span) callback, {
+    Map<String, SentryAttribute>? attributes,
+    SentrySpanV2? parentSpan = const UnsetSentrySpanV2(),
+    DateTime? startTimestamp,
+  }) {
+    throw UnimplementedError();
+  }
+}
+
+class MockScope extends Scope {
+  MockScope() : super(SentryOptions());
+
+  @override
+  SentryUser? user;
+  @override
+  final Map<String, String> tags = {};
+  final Map<String, dynamic> context = {};
+
+  @override
+  Future<void> setUser(SentryUser? user) async {
+    this.user = user;
+  }
+
+  @override
+  Future<void> setTag(String key, String value) async {
+    tags[key] = value;
+  }
+
+  @override
+  Future<void> removeTag(String key) async {
+    tags.remove(key);
+  }
+
+  @override
+  Future<void> setContexts(String key, dynamic value) async {
+    context[key] = value;
+  }
+
+  @override
+  Future<void> removeContexts(String key) async {
+    context.remove(key);
+  }
+
+  @override
+  Future<void> clear() async {
+    user = null;
+    tags.clear();
+    context.clear();
   }
 }